requirements.txt
wheel==0.36.2
Cython==0.29.21
mahotas==1.4.11
matplotlib==3.3.3
Pillow==8.1.0
numpy==1.19.2
scipy==1.5.2
scikit-image==0.18.2
scikit-build==0.11.1
scikit-learn==0.24.0
opencv-python==4.4.0.46
seaborn==0.11.1
imutils==0.5.3
openpyxl==3.0.5
certifi==2020.12.5
imageio==2.9.0
click==7.1.2
pandas==1.1.5
PyYAML==5.3.1
pytest==6.2.1
coverage==5.3.1
coveralls==2.2.0
